# cmd/server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  internal_module_ai_provider_pool.AccountResponse:
    properties:
      consecutive_failures:
        type: integer
      created_at:
        type: string
      daily_limit:
        type: integer
      health_status:
        $ref: '#/definitions/internal_module_ai_provider_pool.HealthStatus'
      id:
        type: string
      is_active:
        type: boolean
      key_prefix:
        type: string
      last_failure_at:
        type: string
      last_health_check:
        type: string
      name:
        type: string
      priority:
        type: integer
      provider_id:
        type: string
      rate_limit_rpm:
        type: integer
      rate_limit_tpm:
        type: integer
      total_cost_usd:
        type: number
      total_requests:
        type: integer
      total_tokens:
        type: integer
      updated_at:
        type: string
      weight:
        type: integer
    type: object
  internal_module_ai_provider_pool.AccountStatsResponse:
    properties:
      account_id:
        type: string
      consecutive_failures:
        type: integer
      daily_stats:
        items:
          $ref: '#/definitions/internal_module_ai_provider_pool.DailyStatsResponse'
        type: array
      health_status:
        $ref: '#/definitions/internal_module_ai_provider_pool.HealthStatus'
      last_health_check:
        type: string
      total_cost_usd:
        type: number
      total_requests:
        type: integer
      total_tokens:
        type: integer
    type: object
  internal_module_ai_provider_pool.CreateAccountRequest:
    properties:
      api_key:
        type: string
      daily_limit:
        type: integer
      name:
        type: string
      priority:
        type: integer
      rate_limit_rpm:
        type: integer
      rate_limit_tpm:
        type: integer
      weight:
        type: integer
    required:
    - api_key
    - name
    type: object
  internal_module_ai_provider_pool.DailyStatsResponse:
    properties:
      cost_usd:
        type: number
      date:
        type: string
      requests_count:
        type: integer
      tokens_count:
        type: integer
    type: object
  internal_module_ai_provider_pool.HealthStatus:
    enum:
    - healthy
    - degraded
    - unhealthy
    type: string
    x-enum-varnames:
    - HealthStatusHealthy
    - HealthStatusDegraded
    - HealthStatusUnhealthy
  internal_module_ai_provider_pool.ListAccountsResponse:
    properties:
      accounts:
        items:
          $ref: '#/definitions/internal_module_ai_provider_pool.AccountResponse'
        type: array
      total:
        type: integer
    type: object
  internal_module_ai_provider_pool.UpdateAccountRequest:
    properties:
      daily_limit:
        type: integer
      is_active:
        type: boolean
      name:
        type: string
      priority:
        type: integer
      rate_limit_rpm:
        type: integer
      rate_limit_tpm:
        type: integer
      weight:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@uniedit.io
    name: UniEdit Support
    url: https://uniedit.io/support
  description: UniEdit 视频编辑器后端 API
  license:
    name: Proprietary
    url: https://uniedit.io/license
  termsOfService: https://uniedit.io/terms
  title: UniEdit Server API
  version: "1.0"
paths:
...
  /admin/providers/{provider_id}/accounts:
    get:
      description: Get all accounts for an AI provider
      parameters:
      - description: Provider ID
        in: path
        name: provider_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_module_ai_provider_pool.ListAccountsResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List provider accounts
      tags:
      - AI
    post:
      consumes:
      - application/json
      description: Add a new account to an AI provider's pool
      parameters:
      - description: Provider ID
        in: path
        name: provider_id
        required: true
        type: string
      - description: Account data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_module_ai_provider_pool.CreateAccountRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_module_ai_provider_pool.AccountResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Add provider account
      tags:
      - AI
  /admin/providers/{provider_id}/accounts/{account_id}:
    delete:
      description: Remove an account from a provider's pool
      parameters:
      - description: Provider ID
        in: path
        name: provider_id
        required: true
        type: string
      - description: Account ID
        in: path
        name: account_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete provider account
      tags:
      - AI
    get:
      description: Get details of a specific provider account
      parameters:
      - description: Provider ID
        in: path
        name: provider_id
        required: true
        type: string
      - description: Account ID
        in: path
        name: account_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_module_ai_provider_pool.AccountResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get provider account
      tags:
      - AI
    patch:
      consumes:
      - application/json
      description: Update a provider account's settings
      parameters:
      - description: Provider ID
        in: path
        name: provider_id
        required: true
        type: string
      - description: Account ID
        in: path
        name: account_id
        required: true
        type: string
      - description: Update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_module_ai_provider_pool.UpdateAccountRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_module_ai_provider_pool.AccountResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update provider account
      tags:
      - AI
...
  /admin/providers/{provider_id}/accounts/{account_id}/stats:
    get:
      description: Get usage statistics for a provider account
      parameters:
      - description: Provider ID
        in: path
        name: provider_id
        required: true
        type: string
      - description: Account ID
        in: path
        name: account_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_module_ai_provider_pool.AccountStatsResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get account statistics
      tags:
      - AI
...
securityDefinitions:
  BearerAuth:
    description: 'Bearer token authentication. Format: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"